movie updated

+++ movie/src/main/java/com/zensar/repository/MovieRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.zensar.entity.MovieEntity;
 
-public interface MovieRepo extends JpaRepository<MovieEntity, Long> {
+public interface MovieRepo extends JpaRepository<MovieEntity, Integer> {
     Optional<MovieEntity> findByTitle(String title);
 }
